+++ Recolor/UI/src/mods/SIPColorComponent/SIPColorComponent.tsx
@@ -8,6 +8,8 @@ import { bindValue, trigger, useValue } from "cs2/api";
 import { useLocalization } from "cs2/l10n";
 import {getModule} from "cs2/modding";
 import classNames from "classnames";
+import { useState } from "react";
+import { FocusDisabled } from "cs2/input";
 
 // These contain the coui paths to Unified Icon Library svg assets
 const uilStandard =                          "coui://uil/Standard/";
@@ -20,6 +22,7 @@ const CanPasteColor$ = bindValue<boolean>(mod.id, "CanPasteColor");
 const SingleInstance$ = bindValue<boolean>(mod.id, 'SingleInstance');
 const DisableSingleInstance$ = bindValue<boolean>(mod.id, 'DisableSingleInstance');
 const MatchesVanillaColorSet$ = bindValue<boolean[]>(mod.id, 'MatchesVanillaColorSet');
+const ShowHexaDecimals$ = bindValue<boolean>(mod.id, 'ShowHexaDecimals');
 
 function copyColor(color : Color) {
     // This triggers an event on C# side and C# designates the method to implement.
@@ -36,9 +39,43 @@ const ColorFieldTheme: Theme | any = getModule(
     "classes"
 )
 
+export const StringInputField : any = getModule(
+    "game-ui/editor/widgets/fields/string-input-field.tsx",
+    "StringInputField"
+)
+
+export const StringInputFieldStyle : Theme | any = getModule(
+    "game-ui/debug/widgets/fields/input-field/input-field.module.scss",
+    "classes"
+)
+
 function changeColor(channel : number, newColor : Color) {
     // This triggers an event on C# side and C# designates the method to implement.
     trigger(mod.id, "ChangeColor", channel, newColor);
+}
+
+export function convertColorToHexaDecimal(color: Color ) : string {
+    const r = Math.round(color.r * 255).toString(16);
+    const g = Math.round(color.g * 255).toString(16);
+    const b = Math.round(color.b * 255).toString(16);
+    const a = Math.round(color.a * 255).toString(16);
+    return "#"+r+g+b+a;
+}
+
+export function convertHexaDecimalToColor(input: string) : Color 
+{
+    let r:number = Math.min(Math.max(parseInt(input.slice(1,3), 16) / 255, 0), 1);
+    let g:number = Math.min(Math.max(parseInt(input.slice(3,5), 16) / 255, 0), 1);
+    let b:number = Math.min(Math.max(parseInt(input.slice(5,7), 16) / 255, 0), 1);
+    let a:number = Math.min(Math.max(parseInt(input.slice(7,9), 16) / 255, 0), 1);
+
+    let color : Color = {
+        r: r,
+        g: g,
+        b: b,
+        a: a,
+    };
+    return color;
 }
 
 export const SIPColorComponent = (props : { channel : number }) => {
@@ -48,7 +85,28 @@ export const SIPColorComponent = (props : { channel : number }) => {
     const CanPasteColor = useValue(CanPasteColor$);    
     const SingleInstance = useValue(SingleInstance$);
     const DisableSingleInstance = useValue(DisableSingleInstance$);
-    const MatchesVanillaColorSet : boolean[] = useValue(MatchesVanillaColorSet$);        
+    const MatchesVanillaColorSet : boolean[] = useValue(MatchesVanillaColorSet$);
+    const ShowHexaDecimals = useValue(ShowHexaDecimals$);
+    
+    let [textInput, setTextInput] = useState(convertColorToHexaDecimal(CurrentColorSet.Channels[props.channel]));
+    let [validInput, setValidInput] = useState(true);
+
+    function HandleTextInput () {
+        if (textInput.length == 9 &&  /^#[0-9A-F]{6}[0-9a-f]{0,2}$/i.test(textInput)) 
+        { 
+            changeColor(props.channel, convertHexaDecimalToColor(textInput));
+            setValidInput(true)
+        }
+        else if (textInput.length == 7 && /^#[0-9A-F]{6}[0-9a-f]{0,2}$/i.test(textInput+"ff")) 
+        {
+            changeColor(props.channel, convertHexaDecimalToColor(textInput+"ff"));      
+            setTextInput(textInput+"ff");      
+            setValidInput(true);
+        } else 
+        {
+            setValidInput(false);          
+        } 
+    }
 
     return (
         <div className={styles.columnGroup}>
@@ -57,7 +115,7 @@ export const SIPColorComponent = (props : { channel : number }) => {
                     className={classNames(ColorFieldTheme.colorField, styles.rcColorField)} 
                     value={CurrentColorSet.Channels[props.channel]} 
                     focusKey={VanillaComponentResolver.instance.FOCUS_DISABLED} 
-                    onChange={(e) => {changeColor(props.channel, e);}}
+                    onChange={(e) => {changeColor(props.channel, e); setTextInput(convertColorToHexaDecimal(e))}}
                     alpha={1}
                 />
             </div>
@@ -94,6 +152,21 @@ export const SIPColorComponent = (props : { channel : number }) => {
                     />
                 )}
             </div>
+            { ShowHexaDecimals && (
+                <FocusDisabled disabled={true}>
+                    <div className={styles.rowGroup}>
+                        <StringInputField
+                            value={textInput.replace(/[\r\n]+/gm, '')}
+                            disabled ={false}
+                            onChange={ (e : string) => { setTextInput(e); }}
+                            onChangeEnd={HandleTextInput}
+                            className={validInput?  classNames(StringInputFieldStyle.textInput, styles.rcColorFieldInput) : classNames(StringInputFieldStyle.textInput, styles.rcColorFieldInput, styles.invalidFieldInput)}
+                            multiline={false}
+                            maxLength={9}
+                        />
+                    </div>
+                </FocusDisabled>
+            )}
         </div>
     );
 }