+++ Recolor/ChangeLog.md
@@ -2,4 +2,6 @@
 * You can select NetLane fences and hedges placed by EDT or inside buildings using a tool activated by keyboard shortcut.
 * Color picker and painter can select Netlane fences and hedges.
 * When changing colors, SubObject and SubLanes colors will update immediately.
-* Color painter can toggle channels
+* Color painter can toggle channels.
+* Alpha transparancy sliders added.
+* Optional Hexadecimal input fields.